Move Transform library contents into Common as well.
Again, should reduce number of libraries to distribute.
