--- src/osvr/Transform/CMakeLists.txt
@@ -1,35 +1,0 @@
-osvr_setup_lib_vars(Transform)
-
-set(API
-    "${HEADER_LOCATION}/ChangeOfBasis.h"
-    "${HEADER_LOCATION}/JSONTransformVisitor.h"
-    "${HEADER_LOCATION}/Transform.h"
-    "${HEADER_LOCATION}/DegreesToRadians.h")
-
-set(SOURCE
-    JSONTransformVisitor.cpp)
-
-osvr_add_library()
-
-set_target_properties(${LIBNAME_FULL} PROPERTIES
-    VERSION ${OSVR_VERSION}
-    SOVERSION 0
-    INTERFACE_osvrcore_MAJOR_VERSION 0)
-
-set_property(TARGET ${LIBNAME_FULL} APPEND PROPERTY
-    COMPATIBLE_INTERFACE_STRING osvrcore_MAJOR_VERSION)
-
-target_include_directories(${LIBNAME_FULL}
-    PUBLIC
-    ${Boost_INCLUDE_DIRS})
-
-target_compile_options(${LIBNAME_FULL}
-    PUBLIC
-    ${OSVR_CXX11_FLAGS})
-
-target_link_libraries(${LIBNAME_FULL}
-    PUBLIC
-    jsoncpp_lib
-    PRIVATE
-    eigen-headers)
-
